refactor(python): group support lib into _support submodule
