+++ python/src/nuri/_support/__init__.py
@@ -1,0 +1,4 @@
+#
+# Project NuriKit - Copyright 2026 SNU Compbio Lab.
+# SPDX-License-Identifier: Apache-2.0
+#
